backend/requirements.txt
django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.11
celery==5.6.2
redis==7.1.0
requests==2.31.0

